:truck: Move action visibility logic to service
